\name{A0SD.LRR}
\alias{A0SD.LRR}
\docType{data}
\title{
Copy number information of selected chromosomes for TCGA sample A1-A0SD.
}
\description{
This is partial level 2 bi-allele copy number data selected from The Cancer Genome Atlas Breast Carcinoma 800K Genome-wide SNP 6.0 array for sample A1-A0SD in the format of DNA copy number.
}
\usage{data(A0SD.LRR)}
\references{
The Cancer Genome Atlas Network, Comprehensive molecular portraits of human breast tumours, _Nature_, 2012
}