\name{CHAT-package}
\alias{CHAT-package}
\alias{CHAT}
\docType{package}
\title{
Clonal Heterogeneity Analysis Tool
}
\description{
CHAT is a collection of methods developed for researchers in the field of cancer genomics to study the intra-heterogeneity of tumors. It is applied on both high-density SNP array data and next generation sequencing data. CHAT estimates tumor purity, ploidy on the genome-wide level. It can also assign each somatic CNA or mutation its fraction of cancer cell carriers, namely, sAGP or CCF. 
}
\details{
\tabular{ll}{
Package: \tab CHAT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-10\cr
License: \tab GPL(>=2) \cr
}
}
\author{
Bo Li <libo@umich.edu>
}
\references{
Li et al., Genomic estimates of aneuploid content in Glioblastoma Multiforme and imporved classification. _Clinical Cancer Research_, 2012.

Li et al., A general framework for characterizing tumor subclonality using DNA sequencing and SNP array data. 2014
}
\keyword{ package }
