% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat}
\alias{chat}
\title{chat}
\usage{
chat(message, temperature, model)
}
\arguments{
\item{message}{The message sent to OpenAI}

\item{temperature}{The sampling temperature, between 0 and 1. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic. If set to 0, the model will use log probability to automatically increase the temperature until certain thresholds are hit. (https://platform.openai.com/docs/api-reference/audio)}

\item{model}{OpenAI's API models}
}
\value{
A dataframe containing the chosen message, temperature, model and the response from OpenAI's API models
}
\description{
chat
}
\examples{
chat("What is an apple", 0.5, "gpt-3.5-turbo-0613")
}
