% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{inclusion_decision_description}
\alias{inclusion_decision_description}
\title{Generate the inclusion decision description}
\usage{
inclusion_decision_description()
}
\value{
A string that describes the decision-making process.
}
\description{
This function generates a string that describes the decision-making process for including a study in the review.
The string is used as a message to the OpenAI API.
}
