\name{prior}
\alias{prior}
\docType{data}
\title{
Prior parameters for MCMC
}
\description{
A pre-selected set of prior parameters required to run \code{DPdensity()}
}
\usage{data(prior)}
\format{
A list containing 7 parameters.
$ a0 : real 0.1
$ b0 : real 0.1
$ m1: real 0.5
$ psiinv1: real 0.01
$ nu1: real 4
$ tau1: real 1
$ tau2: real 100
}