% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.r
\name{set_api_key}
\alias{set_api_key}
\title{Set API Key}
\usage{
set_api_key()
}
\description{
This function sets the API key for the CHATGPT_KEY environment variable.
If the environment variable is not set, it prompts the user to enter the API key.
}
\details{
The function first checks if the "CHATGPT_KEY" environment variable is set.
If it is not set, it prompts the user to enter the value for "CHATGPT_KEY" and sets the environment variable to the entered value.
}
\examples{
set_api_key()

}
