% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.r
\name{set_api_key}
\alias{set_api_key}
\title{Set API Key}
\usage{
set_api_key(key = NULL, env_var = "CHATGPT_KEY")
}
\arguments{
\item{key}{The API key to be set. If not provided, the user will be prompted to enter it.}

\item{env_var}{The environment variable to set the API key to. Default is "CHATGPT_KEY".}
}
\description{
This function sets the API key for the CHATGPT_KEY environment variable.
If no key is provided and the environment variable is not set, it prompts the user to enter the API key.
}
\examples{
set_api_key("your_api_key_here")
set_api_key("your_api_key_here", "MY_ENV_VAR")

}
